PR fixes: rename file, remove duplicate check,
